moves main into its own package and updates deps
This moves the main class into cmd/car so that we can expose an API for
use in other libraries. This also updates dependencies.

Note 1: This doesn't update urlfave as I think we can just use normal go
flags instead of that, which simplifies downstream usage.

Note 2: docker/distribution seems to not really be interested in
releasing a new version ever. A follow-up should probably remove that
dep.

Signed-off-by: Adrian Cole <adrian@tetrate.io>